Two conflicts, both from #897 extracting SharpMUSH.SocketServer out of
SharpMUSH.ConnectionServer.

SharpMUSH.ConnectionServer.csproj: took main's. Every PackageReference this
branch touched moved to SharpMUSH.SocketServer along with the telnet stack, so
$(TelnetNegotiationCoreVersion) is applied there instead. All three references
to the package -- Library, Server, SocketServer -- still resolve through the
one property.

PuebloMxpIntegrationTests.cs: took main's rewrite, which drops the dependency
on the login screen and room contents in favour of publishing known markup
through NATS and waiting on the negotiated format. Re-applied this branch's two
additions in that idiom: MxpNegotiation_ClientAccepts_ServerSendsStartMarker,
and the marker read that MxpNegotiation_ClientAccepts_OutputIncludesMxpTags now
does before asserting, since everything it checks is only true of a client that
entered MXP mode.

SharpMUSH.Tests 6504 passing, SharpMUSH.Tests.BUnit 625 passing, 0 failed.

+++ SharpMUSH.Implementation/Functions/ConnectionFunctions.cs
@@ -106,7 +106,7 @@ public partial class Functions
 		var executor = await parser.CurrentState.KnownExecutorObject(Mediator);
 		var arg0 = parser.CurrentState.Arguments["0"].Message!.ToPlainText();
 
-		if (int.TryParse(arg0, out var port))
+		if (long.TryParse(arg0, out var port))
 		{
 			var data2 = ConnectionService.Get(port);
 			if (data2 is null || data2.Ref is null)
@@ -427,7 +427,7 @@ public partial class Functions
 		var executor = await parser.CurrentState.KnownExecutorObject(Mediator);
 		var arg0 = parser.CurrentState.Arguments["0"].Message!.ToPlainText();
 
-		if (int.TryParse(arg0, out var port))
+		if (long.TryParse(arg0, out var port))
 		{
 			var data2 = ConnectionService.Get(port);
 			if (data2 is null || data2.Ref is null)
